+++ app/src/main/java/com/devmob/alaya/MainContent.kt
@@ -11,8 +11,8 @@ import com.devmob.alaya.domain.model.IconType
 import com.devmob.alaya.domain.model.ItemMenu
 import com.devmob.alaya.ui.screen.HomeScreen
 import com.devmob.alaya.ui.components.BottomBarNavigation
-import com.devmob.alaya.ui.screen.CrisisHandlingScreen
-import com.devmob.alaya.ui.screen.CrisisHandlingViewModel
+import com.devmob.alaya.ui.screen.crisis_handling.CrisisHandlingScreen
+import com.devmob.alaya.ui.screen.crisis_handling.CrisisHandlingViewModel
 import com.devmob.alaya.utils.NavUtils
 
 @Composable
@@ -27,7 +27,7 @@ fun MainContent(navController: NavHostController) {
                     items = listOf(
                         ItemMenu(iconType = IconType.MENU, route = "menu", contentDescription = "menu", order = 3),
                         ItemMenu(iconType = IconType.PATIENT, route = NavUtils.Routes.Crisis.route, contentDescription = "boton para el manejo de crisis", order = 2),
-                        ItemMenu(iconType = IconType.HOME, route = "inicio", contentDescription = "boton de inicio", order = 1),
+                        ItemMenu(iconType = IconType.HOME, route = NavUtils.Routes.Home.route, contentDescription = "boton de inicio", order = 1),
                     ),
                     navHostController = navController
                 )
