+++ app/src/main/java/com/devmob/alaya/ui/screen/crisis_handling/CrisisHandlingViewModel.kt
@@ -1,4 +1,4 @@
-package com.devmob.alaya.ui.screen
+package com.devmob.alaya.ui.screen.crisis_handling
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableIntStateOf
